Change the API path of app-wise token revocation API
