+++ components/org.wso2.carbon.identity.api.user.authorized.apps/org.wso2.carbon.identity.rest.api.user.authorized.apps.v2/src/gen/java/org/wso2/carbon/identity/rest/api/user/authorized/apps/v2/AuthorizedAppsApi.java
@@ -1,18 +1,20 @@
 /*
-* Copyright (c) 2021, WSO2 Inc. (http://www.wso2.com).
-*
-* Licensed under the Apache License, Version 2.0 (the "License");
-* you may not use this file except in compliance with the License.
-* You may obtain a copy of the License at
-*
-* http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-*/
+ * Copyright (c) 2021, WSO2 Inc. (http://www.wso2.com).
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
 
 package org.wso2.carbon.identity.rest.api.user.authorized.apps.v2;
 
@@ -33,13 +35,13 @@ import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.Response;
 
+@Path("/authorized-apps")
+@Api(description = "The authorized-apps API")
 
-@Path("/applications")
-@Api(description = "The applications API")
-public class ApplicationsApi  {
+public class AuthorizedAppsApi  {
 
     @Autowired
-    private ApplicationsApiService delegate;
+    private AuthorizedAppsApiService delegate;
 
     @Valid
     @DELETE
@@ -53,7 +55,7 @@ public class ApplicationsApi  {
             
         })
     }, tags={ "admin" })
-    @ApiResponses(value = {
+    @ApiResponses(value = { 
         @ApiResponse(code = 204, message = "Item Deleted", response = Void.class),
         @ApiResponse(code = 401, message = "Unauthorized", response = Void.class),
         @ApiResponse(code = 403, message = "Resource Forbidden", response = Void.class),
