+++ components/org.wso2.carbon.identity.api.user.authorized.apps/org.wso2.carbon.identity.rest.api.user.authorized.apps.v2/src/gen/java/org/wso2/carbon/identity/rest/api/user/authorized/apps/v2/factories/AuthorizedAppsApiServiceFactory.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright (c) 2021, WSO2 Inc. (http://www.wso2.com).
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.wso2.carbon.identity.rest.api.user.authorized.apps.v2.factories;
+
+import org.wso2.carbon.identity.rest.api.user.authorized.apps.v2.AuthorizedAppsApiService;
+import org.wso2.carbon.identity.rest.api.user.authorized.apps.v2.impl.AuthorizedAppsApiServiceImpl;
+
+public class AuthorizedAppsApiServiceFactory {
+
+   private final static AuthorizedAppsApiService service = new AuthorizedAppsApiServiceImpl();
+
+   public static AuthorizedAppsApiService getAuthorizedAppsApi() {
+      return service;
+   }
+}
